Use uuid for type for image in `CreateRecipe`
Makes it easier to not get error on editing a recipe because of more
strict types. Ideally I think I want to change the type of the database
column to uuid, so need to create a migration script to fix that. Then I
don't need to convert it back and forth between string an uuid so much